[PLAYER-4607] Move IntentBuilder to utils
